pyproject.toml
[project]
name = "mcp-micropython"
version = "0.1.0"
description = "MCP bridge server for MicroPython (ESP32, RP2040, etc.) via USB serial"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyserial>=3.5",
]

[project.scripts]
mcp-micropython = "mcp_micropython.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_micropython"]

